chore: move snippets to own plugin folder, use external plugin
