+++ src/rich-text/editor.ts
@@ -43,8 +43,6 @@ import { interfaceManagerPlugin } from "../shared/prosemirror-plugins/interface-
 import { IExternalPluginProvider } from "../shared/editor-plugin";
 import { createMenuEntries } from "../shared/menu/index";
 import { createMenuPlugin } from "../shared/menu/plugin";
-import { stackSnippetPasteHandler } from "../shared/plugins/stack-snippets/paste-handler";
-import { StackSnippetView } from "../shared/plugins/stack-snippets/snippet-view";
 
 export interface RichTextOptions extends CommonViewOptions {
     /** Array of LinkPreviewProviders to handle specific link preview urls */
@@ -168,18 +166,6 @@ export class RichTextEditor extends BaseView {
                     },
                     html_block_container: function (node: ProseMirrorNode) {
                         return new HtmlBlockContainer(node);
-                    },
-                    stack_snippet: (
-                        node: ProseMirrorNode,
-                        view: EditorView,
-                        getPos: () => number
-                    ) => {
-                        return new StackSnippetView(
-                            node,
-                            view,
-                            getPos,
-                            this.options.stackSnippet
-                        );
                     },
                     ...this.externalPluginProvider.nodeViews,
                 },
