+++ src/rich-text/schema.ts
@@ -1,6 +1,5 @@
 import { Attrs, MarkSpec, NodeSpec } from "prosemirror-model";
 import { _t } from "../shared/localization";
-import { stackSnippetRichTextNodeSpec } from "../shared/plugins/stack-snippets/schema";
 
 //TODO this relies on Stacks classes, should we abstract?
 
@@ -438,8 +437,6 @@ const nodes: {
             tagType: { default: "tag" },
         },
     },
-
-    ...stackSnippetRichTextNodeSpec,
 };
 
 const marks: {
