+++ src/shared/markdown-parser.ts
@@ -11,10 +11,6 @@ import { stackLanguageComments } from "./markdown-it/stack-language-comments";
 import { tagLinks } from "./markdown-it/tag-link";
 import { tight_list } from "./markdown-it/tight-list";
 import type { CommonmarkParserFeatures } from "./view";
-import {
-    stackSnippetMarkdownParser,
-    stackSnippetPlugin,
-} from "./plugins/stack-snippets/schema";
 
 // extend the default markdown parser's tokens and add our own
 const customMarkdownParserTokens: MarkdownParser["tokens"] = {
@@ -154,8 +150,6 @@ const customMarkdownParserTokens: MarkdownParser["tokens"] = {
     code_inline_split: {
         mark: "code",
     },
-
-    ...stackSnippetMarkdownParser,
 };
 
 // add tag attribute support to all the marks like we did in schema
@@ -325,9 +319,6 @@ export function createDefaultMarkdownItInstance(
 
     // ensure we can tell the difference between the different types of hardbreaks
     defaultMarkdownItInstance.use(hardbreak_markup);
-
-    // parse Stack Snippet codes
-    defaultMarkdownItInstance.use(stackSnippetPlugin);
 
     // TODO should always exist, so remove the check once the param is made non-optional
     externalPluginProvider?.alterMarkdownIt(defaultMarkdownItInstance);
