+++ src/shared/menu/entries.ts
@@ -58,7 +58,6 @@ import {
     makeMenuLinkEntry,
     MenuBlock,
 } from "./helpers";
-import { openSnippetModal } from "../plugins/stack-snippets/commands";
 
 /**
  * Creates a dropdown menu for table edit functionality
@@ -373,31 +372,7 @@ export const createMenuEntries = (
                     },
                     "code-block-btn"
                 ),
-            },
-            addIf(
-                {
-                    key: "openSnippetModal",
-                    richText: {
-                        command: openSnippetModal(options.stackSnippet),
-                    },
-                    commonmark: {
-                        command: openSnippetModal(options.stackSnippet),
-                    },
-                    display: makeMenuButton(
-                        "Play",
-                        {
-                            title: _t("commands.stack_snippet.title", {
-                                shortcut: getShortcut("Mod-9"),
-                            }),
-                            description: _t(
-                                "commands.stack_snippet.description"
-                            ),
-                        },
-                        "stack-snippet-open-btn"
-                    ),
-                },
-                "stackSnippet" in options
-            ),
+            }
         ],
     },
     {
