+++ src/shared/view.ts
@@ -6,7 +6,6 @@ import type { EditorView } from "prosemirror-view";
 import { EditorPlugin } from "./editor-plugin";
 import type { ImageUploadOptions } from "./prosemirror-plugins/image-upload";
 import { setAttributesOnElement, stackOverflowValidateLink } from "./utils";
-import { StackSnippetOptions } from "./plugins/stack-snippets/common";
 
 /** Describes each distinct editor type the StacksEditor handles */
 export enum EditorType {
@@ -45,8 +44,6 @@ export interface CommonViewOptions {
     pluginParentContainer?: (view: EditorView) => Element;
     /** Image uploader options */
     imageUpload?: ImageUploadOptions;
-    /** StackSnippet options */
-    stackSnippet?: StackSnippetOptions;
     /** Externally written plugins to add to the editor */
     editorPlugins?: EditorPlugin[];
 }
